Attach sources and javadoc when deploying
The ImageJ Maven repository is currently the destination for
SciJava artifacts, even those which are not ImageJ-specific.
(Some day, it would be nice to rebrand that repository as the
SciJava Maven repository, but for technical reasons that is not
trivial at the moment.)

The deploy-to-imagej profile enables the configuration necessary
for deploying to this repository. Profiles to deploy to other
repositories which become associated with the SciJava ecosystem
are equally feasible.

This change ensures that whenever an artifact is deployed to the
ImageJ Maven repository, the sources and javadoc archives are
also generated and deployed.

This configuration block migrated from sc.fiji:pom-fiji.