Add imagej-maven-plugin configuration
When the imagej.app.directory property is set, the imagej-maven-plugin
will copy the resultant artifact and its dependencies into an ImageJ
application installation, removing old versions as needed.

Hopefully some day this logic will be generalized and migrate upstream
into scijava-maven-plugin, since really any SciJava-based application
could benefit from this feature. But at the moment it is specific to
the imagej-maven-plugin, and that is good enough for now.

This configuration block migrated from sc.fiji:pom-fiji.