Use banDuplicateClasses of extra-enforcer-rules
It used to be that the banDuplicateClasses rule of the Extra Enforcer
Rules had a bug (mojohaus/extra-enforcer-rules#7) which caused erroneous
detection of duplicate classes across multiple Vs of the same GA.

However, happily, this bug is now fixed, and extra-enforcer-rules
version 1.0-beta-5 and newer contain the fix. Therefore, we no longer
need to use our fork, and can instead use the upstream rule as desired.