Make another attempt to fix sources double deploy
Firstly, the maven-source-plugin in the sonatype-oss-release and
deploy-to-imagej profiles was redundant because it replicated what
is already enabled by default. So those sections have been removed.

Secondly, using the jar-no-fork goal still results in double
generation for reasons I don't understand. This is discussed here:

    https://issues.apache.org/jira/browse/MNG-5939

And some people had success using the jar goal instead. So let's try it.