Let projects define a minimum Java build version
By default, we require at least Java 5. In practice, a later version
than 1.5 will be needed, depending on the version of Maven used;
e.g., Maven 3.3.1+ requires Java 7 or later. We are just setting a
conservative explicit default here.

The main point of this commit is to allow downstream projects to
override with the in-practice required minimum, which might even
include a sub-version. This lets subprojects give more meaningful error
messages when working around bugs in Java which have since been fixed.

This commit is dedicated to Juan Nunez-Iglesias; see:
https://github.com/imglib/imglib2/issues/192#issuecomment-349212095