Update the minimum Maven version to 3.3.9
Older versions fail with the following error message:

  Unable to parse configuration of mojo org.apache.maven.plugins:maven-enforcer-plugin:1.4.1:enforce for parameter ignoreClasses: Cannot assign configuration entry 'ignoreClasses' with value 'org.hibernate.stat.ConcurrentStatisticsImpl,org.junit.runner.Runner' of type java.lang.String to property of type java.lang.String[] -> [Help 1]

This is because the ignoreClasses configuration of maven-enforcer-plugin
is an array, which classically was specified with a list of ignoreClass
elements. But as of 3.3.9, you can set it directly to a string of
comma-separated values, and Maven magically converts it. This is a great
feature because now overridding the ignoreClasses configuration can be
done via a property, where previously it was necessary to include a
bulky <build><plugins><plugin>... section.

While duplicate classes in a project's dependency set are strongly to be
avoided, it is infeasible in some scenarios. For example, the JMockit
library replicates part of JUnit by design, and so an exclusion must
be given in order to make use of this library at all.

In conclusion, it is worth it to bump the minimum Maven version in order
to support this. Some package managers may ship a too-old version now,
but installing Maven in user-space is as easy as unpacking a ZIP, so
it's not the end of the world.