Make a guess at a reasonable package prefix
Our goal here is to choose a likely-to-be-unique, descriptive and legal
name, Not to glean what the actual package prefix is for the project.

This is important because the package-name is also used as the
automatic-module-name value by default.

Unfortunately, it seems regex-property cannot operate on properties
defined in the <properties> section -- only those basic properties of
the project such as project.groupId and project.artifactId.