Remove jigsaw profile
Unfortunately, this profile does not work if you are compiling a
modularized project targeting Java 9+, using Java 9+. It only works
if you are targeting Java 8 or earlier, but compiling with Java 9+.
This is an exceedingly rare case (I didn't find any SciJava ecosystem
projects doing it), and we need to support compilation of modularized
Java 11 projects out of the box. Therefore, we remove this profile, to
avoid breakage. For modularized projects needing to support Java 8 or
earlier, they will need to add the maven-compiler-plugin configuration
directly to their own POMs manually, as described at:

  https://maven.apache.org/plugins/maven-compiler-plugin/examples/module-info.html