Use --release flag to target intended JDK version
Starting with Java 9, the --release flag makes it possible to
compile Java code with a given version of the JDK but targeting an
earlier version of the Java API. The --source and --target flags alone
are not enough to guarantee this; the bootclasspath must also be set
appropriately, which the --release flag takes care of doing.

This change passes the value of the scijava.jvm.version property for
--release, as long as the version of Java being used is 9 or later.

With this enhancement, it is no longer necessary to restrict the
build environment (scijava.jvm.build.version) to Java 8 only.

This change could potentially be breaking for projects that override
scijava.jvm.version with a version like 1.6, because the --release flag
does not accept the 1.x syntax, only the new-style 6, 7, 8, etc.
Therefore, we bump the major digit, to indicate that adjustment may be
necessary in downstream POMs.

Co-authored-by: Curtis Rueden <ctrueden@wisc.edu>