Introduce scijava.platform.arch property
And get rid of scijava.platform.bits, which is rather useless.

This property defaults to the same value as os.arch, except that it
tries to unify synonyms, according to the native-lib-loader scheme.
Notably, this scheme collapses x86 32-bit and x86 64-bit platforms to
simply "32" and "64" respectively.

While I'm not convinced that the native-lib-loader scheme is a *good*
scheme, it is what we are using in practice, at least for projects such
as flimlib:flimlib.

Closes #30.