CI: rename env vars for Central credentials
I am not certain whether existing user tokens were migrated over,
but my initial test did not look promising. In the interest of clarity,
this change renames the variable prefix from OSSRH_ to CENTRAL_.

On the scijava-scripts side, the ci-build.sh script can look for the
correct environment variables, and fail fast with a nice message when
the old variables are still set.